Report back pbap and map uuids to AdapterProperties in Java
It seems services were supposed to be enabled by {@link
btif_enable_service}, which would also automatically update the {@link
btif_enabled_services} mask. This "enabled services mask" is the basis
of the local UUIDs that is reported back up to Java (e.g., {@link
AdapterProperties.getUuids}).

However, it seems like only the HFP (server/client) and A2DP
(source/sink) profiles go through this codepath, but not PBAP or MAP
(server/client), which is why their UUIDs are not reported back to Java.

This CL adds calls to {@link btif_enable_service} when the SDP records
for PBAP/MAP (server/client) are created in order to update the "enabled
services mask". This is analogous to how their UUIDs are added to the
EIR when their SDP records are created.

Because these services were enabled elsewhere, the codepath following
the call to {@link btif_enable_service} needs to be modified to catch
the respective service IDs and no-op (instead of enabling a service).
Similar, the "enabled services mask" is updated when the respective SDP
record is removed.

Tag: #stability
Bug: 204597459
Test: atest BluetoothInstrumentationTests
Change-Id: I4c4f3db01c18b2db674ec639c2b714a3c4fec81b